requirements.txt
pip==22.0.4
setuptools==62.0.0
wheel==0.37.1
python-dotenv~=0.20.0
Flask~=2.1.1